// plugins/catalog/src/components/CatalogPage/types.ts
/*
 * Copyright 2020 Spotify AB
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { Entity } from '@backstage/catalog-model';

export type FormValues = {
  kind?: string;
  type?: string[];
};

export type BasicEntry = { id: string; label: string };

export type FilterOptions = {
  kind: BasicEntry[];
  type?: BasicEntry[] | undefined;
};

export type FilterOutput = {
  addServerFilter(key: string, value: string): void;
  addClientFilter(filter: (entity: Entity) => boolean): void;
};

export type FilterSpec =
  | {
      name: string;
      type: 'select';
      defaultValue?: string;
      options(formValues: FormValues, entities: Entity[]): string[];
      outputs(value: string, output: FilterOutput): void;
    }
  | {
      name: string;
      type: 'multiselect';
      defaultValue?: string[];
      options(formValues: FormValues, entities: Entity[]): string[];
      outputs(value: string[], output: FilterOutput): void;
    };